Add libisyntax_get_barcode()
This adds a way to retrieve the slide barcode. (Only works if the barcode information has not been deleted/blanked out.)

This changes the signature of libisyntax_open(), replacing is_init_allocators with an enum for flags (with the flag LIBISYNTAX_OPEN_FLAG_INIT_ALLOCATORS=1 emulating the old behavior for compatiblity).